+++ classes/output/core/course_renderer.php
@@ -27,6 +27,9 @@ namespace theme_foundation\output\core;
 
 defined('MOODLE_INTERNAL') || die();
 
+use html_writer;
+use moodle_url;
+
 /**
  * The course renderer.
  *
@@ -42,5 +45,40 @@ class course_renderer extends \core_course_renderer {
     protected function get_mustache() {
         $toolbox = \theme_foundation\toolbox::get_instance();
         return $toolbox->get_mustache();
+    }
+
+    /**
+     * Returns HTML to display course name.
+     *
+     * @param coursecat_helper $chelper
+     * @param core_course_list_element $course
+     * @return string
+     */
+    protected function course_name(\coursecat_helper $chelper, \core_course_list_element $course): string {
+        $content = '';
+        if ($chelper->get_show_courses() >= self::COURSECAT_SHOW_COURSES_EXPANDED) {
+            $nametag = 'h3';
+        } else {
+            $nametag = 'div';
+        }
+        $toolbox = \theme_foundation\toolbox::get_instance();
+        $coursename = $toolbox->getfontawesomemarkup('graduation-cap', array('mr-1')).$chelper->get_course_formatted_name($course);
+        $coursenamelink = html_writer::link(new moodle_url('/course/view.php', ['id' => $course->id]),
+            $coursename, ['class' => $course->visible ? 'aalink' : 'aalink dimmed']);
+        $content .= html_writer::tag($nametag, $coursenamelink, ['class' => 'coursename']);
+        // If we display course in collapsed form but the course has summary or course contacts, display the link to the info page.
+        $content .= html_writer::start_tag('div', ['class' => 'moreinfo']);
+        if ($chelper->get_show_courses() < self::COURSECAT_SHOW_COURSES_EXPANDED) {
+            if ($course->has_summary() || $course->has_course_contacts() || $course->has_course_overviewfiles()
+                || $course->has_custom_fields()) {
+                $url = new moodle_url('/course/info.php', ['id' => $course->id]);
+                $image = $this->output->pix_icon('i/info', $this->strings->summary);
+                $content .= html_writer::link($url, $image, ['title' => $this->strings->summary]);
+                // Make sure JS file to expand course content is included.
+                $this->coursecat_include_js();
+            }
+        }
+        $content .= html_writer::end_tag('div');
+        return $content;
     }
 }